# docs/swagger.yaml
basePath: /api/v1
definitions:
  github_com_Tencent_WeKnora_internal_errors.AppError:
    properties:
      code:
        $ref: '#/definitions/github_com_Tencent_WeKnora_internal_errors.ErrorCode'
      details: {}
      message:
        type: string
    type: object
  github_com_Tencent_WeKnora_internal_errors.ErrorCode:
    enum:
    - 1000
    - 1001
    - 1002
    - 1003
    - 1004
    - 1005
    - 1006
    - 1007
    - 1008
    - 1009
    - 1010
    - 2000
    - 2001
    - 2002
    - 2003
    - 2004
    - 2100
    - 2101
    - 2102
    - 2103
    type: integer
    x-enum-varnames:
    - ErrBadRequest
    - ErrUnauthorized
    - ErrForbidden
    - ErrNotFound
    - ErrMethodNotAllowed
    - ErrConflict
    - ErrTooManyRequests
    - ErrInternalServer
    - ErrServiceUnavailable
    - ErrTimeout
    - ErrValidation
    - ErrTenantNotFound
    - ErrTenantAlreadyExists
    - ErrTenantInactive
    - ErrTenantNameRequired
    - ErrTenantInvalidStatus
    - ErrAgentMissingThinkingModel
    - ErrAgentMissingAllowedTools
    - ErrAgentInvalidMaxIterations
    - ErrAgentInvalidTemperature
...
  internal_handler.UpdateChunkRequest:
    properties:
      chunk_index:
        type: integer
      content:
        type: string
      embedding:
        items:
          type: number
        type: array
      end_at:
        type: integer
      image_info:
        type: string
      is_enabled:
        type: boolean
      start_at:
        type: integer
    type: object
...
info:
  contact:
    name: WeKnora Github
    url: https://github.com/aiplusall/aiplusall-kb
  description: WeKnora 知识库管理系统 API 文档
  termsOfService: http://swagger.io/terms/
  title: WeKnora API
  version: "1.0"
paths:
...
  /chunks/{knowledge_id}:
    delete:
      consumes:
      - application/json
      description: 删除指定知识下的所有分块
      parameters:
      - description: 知识ID
        in: path
        name: knowledge_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_Tencent_WeKnora_internal_errors.AppError'
      security:
      - Bearer: []
      - ApiKeyAuth: []
      summary: 删除知识下所有分块
      tags:
      - 分块管理
    get:
      consumes:
      - application/json
      description: 获取指定知识下的所有分块列表，支持分页
      parameters:
      - description: 知识ID
        in: path
        name: knowledge_id
        required: true
        type: string
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 分块列表
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_Tencent_WeKnora_internal_errors.AppError'
      security:
      - Bearer: []
      - ApiKeyAuth: []
      summary: 获取知识分块列表
      tags:
      - 分块管理
  /chunks/{knowledge_id}/{id}:
    delete:
      consumes:
      - application/json
      description: 删除指定的分块
      parameters:
      - description: 知识ID
        in: path
        name: knowledge_id
        required: true
        type: string
      - description: 分块ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_Tencent_WeKnora_internal_errors.AppError'
        "404":
          description: 分块不存在
          schema:
            $ref: '#/definitions/github_com_Tencent_WeKnora_internal_errors.AppError'
      security:
      - Bearer: []
      - ApiKeyAuth: []
      summary: 删除分块
      tags:
      - 分块管理
    put:
      consumes:
      - application/json
      description: 更新指定分块的内容和属性
      parameters:
      - description: 知识ID
        in: path
        name: knowledge_id
        required: true
        type: string
      - description: 分块ID
        in: path
        name: id
        required: true
        type: string
      - description: 更新请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handler.UpdateChunkRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 更新后的分块
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_Tencent_WeKnora_internal_errors.AppError'
        "404":
          description: 分块不存在
          schema:
            $ref: '#/definitions/github_com_Tencent_WeKnora_internal_errors.AppError'
      security:
      - Bearer: []
      - ApiKeyAuth: []
      summary: 更新分块
      tags:
      - 分块管理
  /chunks/by-id/{id}:
    get:
      consumes:
      - application/json
      description: 仅通过分块ID获取分块详情（不需要knowledge_id）
      parameters:
      - description: 分块ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 分块详情
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_Tencent_WeKnora_internal_errors.AppError'
        "404":
          description: 分块不存在
          schema:
            $ref: '#/definitions/github_com_Tencent_WeKnora_internal_errors.AppError'
      security:
      - Bearer: []
      - ApiKeyAuth: []
      summary: 通过ID获取分块
      tags:
      - 分块管理
  /chunks/by-id/{id}/questions:
    delete:
      consumes:
      - application/json
      description: 删除分块中生成的问题
      parameters:
      - description: 分块ID
        in: path
        name: id
        required: true
        type: string
      - description: 问题ID
        in: body
        name: request
        required: true
        schema:
          properties:
            question_id:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_Tencent_WeKnora_internal_errors.AppError'
        "404":
          description: 分块不存在
          schema:
            $ref: '#/definitions/github_com_Tencent_WeKnora_internal_errors.AppError'
      security:
      - Bearer: []
      - ApiKeyAuth: []
      summary: 删除生成的问题
      tags:
      - 分块管理
...
securityDefinitions:
  ApiKeyAuth:
    description: 租户身份认证：输入 sk- 开头的 API Key
    in: header
    name: X-API-Key
    type: apiKey
  Bearer:
    description: 用户登录认证：输入 Bearer {token} 格式的 JWT 令牌
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"